+++ src/modules/esm.mjs
@@ -1,27 +1,26 @@
-import path from 'path';
-import { release, version } from 'os';
-import { createServer } from 'http';
-
-import './files/c.js';
+const path = require('path');
+const { release, version } = require('os');
+const { createServer: createServerHttp } = require('http');
+require('./files/c');
 
 const random = Math.random();
 
 let unknownObject;
 
 if (random > 0.5) {
-    unknownObject = await import('./files/a.json', { with : { type: 'json' } });
+    unknownObject = require('./files/a.json');
 } else {
-    unknownObject = await import('./files/b.json', { with : { type: 'json' } });
+    unknownObject = require('./files/b.json');
 }
 
 console.log(`Release ${release()}`);
 console.log(`Version ${version()}`);
 console.log(`Path segment separator is "${path.sep}"`);
 
-console.log(`Path to current file is ${import.meta.url}`);
-console.log(`Path to current directory is ${path.dirname(import.meta.url)}`);
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
 
-const myServer = createServer((req, res) => {
+const myServer = createServerHttp((_, res) => {
     res.end('Request accepted');
 });
 
@@ -34,4 +33,7 @@ myServer.listen(PORT, () => {
     console.log('To terminate it, use Ctrl+C combination');
 });
 
-export { unknownObject, myServer };
+module.exports = {
+    unknownObject,
+    myServer,
+};
